Major change in how Postscribe is built and structured.
- Now using Webpack for packaging Postscribe as a UMD module
  This means Postscribe is now usable with browser globals, CommonJS, AMD, …
- Replaced JSHint with ESLint
  Does the same thing, but better
- Replaced the `set`-function with object-assign
- Cleaned up code according to stricter ESLint rules
