--- .jshintrc
@@ -1,30 +1,0 @@
-{
-  "browser": {
-    "loopfunc": true,
-    "expr": true,
-    "evil": true,
-    "eqnull": true
-  },
-  "curly": true,
-  "eqeqeq": true,
-  "forin": true,
-  "indent": 2,
-  "latedef": false,
-  "newcap": true,
-  "noarg": true,
-  "node": {
-    "console": false,
-    "process": false,
-    "module": true,
-    "require": false,
-    "__dirname": false,
-    "exports": true
-  },
-  "noempty": true,
-  "quotmark": "single",
-  "sub": true,
-  "trailing": true,
-  "undef": true,
-  "unused": "vars",
-  "white": "vars"
-}
