+++ Gruntfile.js
@@ -1,4 +1,6 @@
-/*jshint node:true*/
+var path = require('path');
+var webpack = require('webpack');
+
 module.exports = function(grunt) {
 
   var pkg = grunt.file.readJSON('package.json');
@@ -13,11 +15,20 @@ module.exports = function(grunt) {
   grunt.initConfig({
     pkg: pkg,
     meta: {
-      banner: '/* <%= pkg.description %>, v<%= pkg.version %> <%= pkg.homepage %>\n' +
-        'Copyright (c) <%= grunt.template.today("yyyy") %> <%= pkg.author.name %>, MIT license ' +
-        '<%= pkg.licenses[0].url %> */'
+      banner: '<%= pkg.description %>, v<%= pkg.version %> <%= pkg.homepage %>\n' +
+        'Copyright (c) <%= grunt.template.today(\'yyyy\') %> <%= pkg.author.name %>, MIT license ' +
+        '<%= pkg.licenses[0].url %>'
     },
-
+    clean: {
+      build: ['dist']
+    },
+    eslint: {
+      options: {
+        configFile: './.eslintrc'
+      },
+      node: ['Gruntfile.js', 'test/generate_expected.phantom.js'],
+      browser: ['src']
+    },
     jshint: {
       // Grr, plugin keeps trying to iterate this...
       jshintrc: './.jshintrc',
@@ -27,27 +38,38 @@ module.exports = function(grunt) {
       },
       // Just for the 'browser' src files
       browser: {
-        src: ['postscribe.js', 'htmlParser/htmlParser.js', 'test/test.js']
+        src: ['src/**/*.js']
       }
     },
-    concat: {
-      options: {
-        banner: '<%= meta.banner %>'
+    webpack: {
+      build: {
+        entry: {
+          postscribe: './src/postscribe.js'
+        },
+        output: {
+          path: path.join(__dirname, 'dist'),
+          filename: 'postscribe.js',
+          library: ['postscribe'],
+          libraryTarget: 'umd'
+        },
+        plugins: [
+          new webpack.BannerPlugin('<%= meta.banner %>')
+        ]
       },
-      dist: {
-        src: ['<%= meta.banner %>', 'htmlParser/htmlParser.js', 'postscribe.js'],
-        dest: 'dist/postscribe.js'
-      }
-    },
-    // Minify postscribe src to postscribe.min.js, prepending a banner
-    uglify: {
-      options: {
-        banner: '<%= meta.banner %>'
-      },
-      dist: {
-        files: {
-          'dist/postscribe.min.js': ['dist/postscribe.js']
-        }
+      buildmin: {
+        entry: {
+          postscribe: './src/postscribe.js'
+        },
+        output: {
+          path: path.join(__dirname, 'dist'),
+          filename: 'postscribe.min.js',
+          library: ['postscribe'],
+          libraryTarget: 'umd'
+        },
+        plugins: [
+          new webpack.BannerPlugin('<%= meta.banner %>'),
+          new webpack.optimize.UglifyJsPlugin()
+        ]
       }
     },
 
@@ -56,11 +78,11 @@ module.exports = function(grunt) {
     },
 
     watch: {
-      files: ['postscribe.js', 'test/*'],
-      tasks: 'jshint qunit'
+      files: ['src/**/*', 'test/*'],
+      tasks: ['webpack:build', 'eslint', 'qunit']
     },
 
-    generate_expected: {
+    generateExpected: {
       dest: 'test/expected.js',
       index: '<%= qunit.files %>',
       phantom: 'test/generate_expected.phantom.js'
@@ -71,7 +93,7 @@ module.exports = function(grunt) {
   grunt.registerTask('generate_expected', 'Generate Files', function() {
     var done = this.async();
 
-    var data = grunt.config('generate_expected');
+    var data = grunt.config('generateExpected');
     var args = [data.phantom, _.first(data.index), data.dest];
     console.info(args);
     grunt.util.spawn({cmd: './node_modules/.bin/phantomjs', args: args}, function(error, result) {
@@ -83,10 +105,10 @@ module.exports = function(grunt) {
     });
   });
 
-  // Alias test
-  grunt.registerTask('test', ['generate_expected', 'qunit']);
-  grunt.registerTask('min', ['concat', 'uglify']);
+  grunt.registerTask('test', ['build', 'generate_expected', 'qunit']);
+  grunt.registerTask('lint', ['eslint']);
+  grunt.registerTask('build', ['clean', 'webpack:build', 'webpack:buildmin']);
   // This is what gets run when you don't specify an argument for grunt.
-  grunt.registerTask('default', ['jshint', 'test']);
+  grunt.registerTask('default', ['lint', 'test']);
 
 };
