--- htmlParser/README.md
@@ -1,6 +1,0 @@
-htmlParser
-==========
-
-A stateless HTML Parser written in JavaScript
-Based on http://ejohn.org/blog/pure-javascript-html-parser/
-Turns an HTML string into a token iterator.
