+++ test/test.html
@@ -2,9 +2,7 @@
 <html>
   <head>
     <!-- source files -->
-    <script src="../postscribe.js"></script>
-    <script src="../htmlParser/htmlParser.js"></script>
-
+    <script src="../dist/postscribe.js"></script>
     <!-- vendor -->
     <script src="vendor/jquery.js"></script>
     <link rel="stylesheet" href="vendor/qunit.css" type="text/css" media="screen" />
